refactor: 항상 repository는 mysql 하위에 셋업되어야 설정 반영됨
